[#93] Splitting XML tests to reflect node file structure
