+++ src/FluentSerializer.Xml.Tests/FluentSerializer.Xml.Tests.csproj
@@ -22,14 +22,6 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-	  <None Remove="Tests\Converting\Converters\EnumConverterTests.Deserialize.Element.md" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <Content Include="Tests\Converting\Converters\EnumConverterTests.Deserialize.Element.md" />
-	</ItemGroup>
-
-	<ItemGroup>
 		<PackageReference Include="coverlet.msbuild" Version="3.1.2">
 			<PrivateAssets>all</PrivateAssets>
 			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
