+++ src/FluentSerializer.Xml.Tests/Tests/DataNodes/Nodes/XmlDocumentTests.Parsing.md
@@ -1,0 +1,4 @@
+# FluentSerializer.Xml.Tests.Tests.DataNodes.Nodes.XmlDocumentTests for Parsing
+
+There currently is no parse functionality for documents
+Parsing of elements jus skip the declaration node
