+++ src/FluentSerializer.Xml.Tests/Tests/DataNodes/Nodes/XmlFragmentTests.Parsing.md
@@ -1,0 +1,3 @@
+# FluentSerializer.Xml.Tests.Tests.DataNodes.Nodes.XmlFragmentTests for Parsing
+
+Xml fragments have no parsing by nature.
